now highaddr

+++ include/sys/types.h
@@ -60,8 +60,8 @@ typedef unsigned int bool;
 #define false	0
 #define true	1
 
-typedef long uintptr_t;
-static_assert(sizeof(void *)==sizeof(long), "error");
+typedef unsigned long uintptr_t;
+static_assert(sizeof(void *)==sizeof(uintptr_t), "error");
 
 typedef uint32_t atomic_t;
 typedef int32_t satomic_t;
